refactored to color picker component
